pyproject.toml
[project]
name = "code-agent-experiments"
version = "0.1.0"
description = "LLM tool-use experiments for code file discovery"
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "Your Name" }]
dependencies = [
    "python-dotenv>=1.1.1",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "ulid>=1.1",
    "openai>=1.0.0",
    "openai-agents>=0.4.0",
    "typer>=0.12",
    "rich>=13",
    "orjson>=3",
    "faiss-cpu>=1.7.0",
    "chromadb>=0.5.0",
    "pgvector>=0.2.0",
    "tree-sitter>=0.23.0",
    "numpy>=1.26",
    "pandas>=2.2",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "uv[dev]>=0.9.5",
    "nox>=2024.4.15",
    "ruff>=0.14.1",
    "pyright>=1.1.406",
    "pip-audit>=2.9.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "freezegun>=1.5.5",
    "pytest-mock>=3.15.1",
    "pytest-xdist>=3.8.0",
    "responses>=0.25.8",
    "pexpect>=4.9.0",
    "pytest-asyncio>=1.2.0",
    "diff-cover>=9.7.1",
    "commitizen>=4.9.1",
    "python-semantic-release>=10.4.1",
    "py-spy>=0.4.1",
    "scalene>=1.5.55",
    "hatch>=1.15.1",
    "pdm>=2.26.0",
    "pre-commit>=4.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
# Install dev dependencies with 'uv sync --extra dev' or 'uv pip install -e .[dev]'

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", "D203", "D213", "D300",
    "I001",
    "ANN401",
    "FBT001", "FBT002",
    "PLR0913",
    "S110",
    "BLE001",
    "PLC0415",
    "TRY400",
    "UP035",
    "EXE001",
    "EXE002",
    "N802",
]
fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "TRY003", "EM101"]
"tests/helpers/pexpect_debug.py" = ["T201", "TRY300"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
include = ["src", "noxfile.py", "tests"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    ".venv",
    "venv",
    ".nox",
    "build",
    "dist",
]
stubPath = "typings"
typeCheckingMode = "strict"
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-v -ra --strict-markers"
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks integration tests",
  "asyncio: marks tests as asyncio"
]
